+++ flutter/lib/app/theme.dart
@@ -1,5 +1,6 @@
 export 'theme/app_alphas.dart';
 export 'theme/app_colors.dart';
+export 'theme/app_motion.dart';
 export 'theme/app_radii.dart';
 export 'theme/app_shadows.dart';
 export 'theme/app_spacing.dart';